Move local_solver to base
