+++ python/dolfinx_eqlb/wrappers.cpp
@@ -16,7 +16,7 @@
 
 #include <dolfinx_eqlb/base/BoundaryData.hpp>
 #include <dolfinx_eqlb/base/FluxBC.hpp>
-#include <dolfinx_eqlb/local_solver.hpp>
+#include <dolfinx_eqlb/base/local_solver.hpp>
 #include <dolfinx_eqlb/reconstruction.hpp>
 
 #include <cstdint>
@@ -55,7 +55,7 @@ void declare_lsolver(py::module& m)
       [](std::vector<std::shared_ptr<dolfinx::fem::Function<T>>>& sol_elmt,
          const dolfinx::fem::Form<T>& a,
          const std::vector<std::shared_ptr<const dolfinx::fem::Form<T>>>& l)
-      { local_solver_lu<T>(sol_elmt, a, l); },
+      { base::local_solver_lu<T>(sol_elmt, a, l); },
       py::arg("solution"), py::arg("a"), py::arg("l"),
       "Local solver based on the LU decomposition");
 
@@ -64,7 +64,7 @@ void declare_lsolver(py::module& m)
       [](std::vector<std::shared_ptr<dolfinx::fem::Function<T>>>& sol_elmt,
          const dolfinx::fem::Form<T>& a,
          const std::vector<std::shared_ptr<const dolfinx::fem::Form<T>>>& l)
-      { local_solver_cholesky<T>(sol_elmt, a, l); },
+      { base::local_solver_cholesky<T>(sol_elmt, a, l); },
       py::arg("solution"), py::arg("a"), py::arg("l"),
       "Local solver based on the Cholesky decomposition");
 
@@ -73,7 +73,7 @@ void declare_lsolver(py::module& m)
       [](std::vector<std::shared_ptr<dolfinx::fem::Function<T>>>& sol_elmt,
          const dolfinx::fem::Form<T>& a,
          const std::vector<std::shared_ptr<const dolfinx::fem::Form<T>>>& l)
-      { local_solver_cg<T>(sol_elmt, a, l); },
+      { base::local_solver_cg<T>(sol_elmt, a, l); },
       py::arg("solution"), py::arg("a"), py::arg("l"),
       "Local solver based on a Conjugated-Gradient solver");
 }
